Rebase on vladimirov/typescript-support2
